Re-add x86_64-linux as a gem platform
I'm not sure what has gone wrong with the gem platform for this but it
has ended up with a gem that doesn't exist (sass-embedded
(1.83.4-x86_64-linux)).

To try rectify this I've run `bundle lock --remove-platform x86_64-linux` and
then `bundle lock --add-platform x86_64-linux` to try sort it out.