Configure app for Heroku deployment
In order to deploy to Heroku, we need to do a few things.

First we need to set the Ruby version in the Gemfile. While it does
technically build on Heroku without this, you get a warning and it'll
build with the latest Ruby version. For the sake of consistency, it's
probably best if it builds with the Ruby version we want (i.e. the one
set in the `.ruby-version` file). So we'll set the Ruby version in the
Gemfile by reading the contents of that file.

Heroku also doesn't like sqlite because the apps are backed by an
ephemeral filesystem which gets cleared periodically[1]. So the database
would get deleted every 24 hours, which clearly isn't ideal. Instead
they push you towards their postgres offering, which we'll use here.

We need to add the `pg` gem to the Gemfile so we can connect to
postgres, then configure the `database.yml` file to use postgres as the
database in production. To keep things as they are, we're still using
sqlite in development.

[1] https://devcenter.heroku.com/articles/sqlite3