Enhanced Request Specs with Devise Integration and Ownership Validation
This commit refactors our test suite to address authentication and ownership issues across quizzes, questions, and quiz taking flows. Previously, request specs were failing due to Devise mapping errors and missing user associations. I updated the specs to create a valid user and associate quizzes with that user. In addition, I integrated Warden’s test helpers (and enabled test mode) so that methods like login_as can be used reliably in our request specs, eliminating the "Could not find a valid mapping" errors.

All tests that require authentication now properly log in a user, while guest flows in the quiz taking specs are simulated by posting to our guest_sign_in route with a guest username. I also added a basic user model spec to ensure our validations (presence and uniqueness for username and email) and Devise functionality are working as expected.

In the DELETE quiz spec, the second quiz was originally defined using 'let', which caused it not to be created and thus made the test fail. Changing it to 'let!' ensured that both quizzes are present when the test runs. To satisfy RuboCop's RSpec/LetSetup rule, I explicitly referenced the second quiz in the test. This fix guarantees that the index view accurately reflects the presence of the second quiz after the first is deleted.